Add a threadsafe sql flag and axotor.py example
This commit adds a threadsafe flag for the sql database.
Threading is useful in chat clients, so that the database
does not need to be written to disk each time it is updated
in threaded applications such as the new axotor.py example.
This commit is backward-compatible with previous applications
so no changes to them are required.

You do need to be sure to lock each thread, so that only
one thread at a time is operating on the database!

This commit also includes a new axotor.py chat script
that communicates through a tor proxy. In addition to the
modules listed in the axotor.py docstring, tor itself must
be installed on the system.