Fix bug by making `mode` local instead of `self.mode`
In a previous version, `self` was an `Axolotl` object. Now that
`SqlitePersistence` handles the database, `mode` should only be a
local variable to be returned by `load_state`. `self.` should have
been deleted along with the other references. The code was only
working because the `mode` returned as a number is successfully
evaluated like a boolean.