Enable `AxolotlConversation` to handle skipped messages on memory
The `staged_hk_mk` dictionary is now composed of `SkippedMessageKeys`
in order to persist both keys and timestamps. This dictionary is
committed by `save_conversation` and loaded by `load_conversation`.
Staging and decryption attempts only have to use this dictionary so
that there is no need to access the persistence for every operation.

The deletion of expired skipped message keys has been moved out of
`commit_skipped_mk` to `delete_expired_skipped_mk`, which is called
every time a persisted `Axolotl` object is loaded.