This commit adds window resizing to two examples
The example programs axochat.py and axotor.py use
a curses interface. Up to now, that interface has not
been resizable. This commit fixes that.

There is a bug in the python curses Textbox class that
gives a stacktrace when the width x height of the Textbox
is greater than the recursion limit. See
https://bugs.python.org/issue13051 for details. By
increasing the recursion limit to 4096, both axochat.py
and axotor.py will now work for windows up to 512
columns in width (the textbox is always 8 rows high).